feat: newsletter signup
