+++ apps/indexer/package.json
@@ -27,7 +27,9 @@
     "viem": "latest",
     "ws": "^8.16.0",
     "axios": "^1.6.8",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "resend": "^3.2.0",
+    "vconsole": "^3.15.1"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.202",
