+++ apps/webapp/components/pages/home/newsletter/index.ts
@@ -1,0 +1,1 @@
+export * from './newsletter-form'
