+++ package.json
@@ -27,8 +27,7 @@
     "@types/express": "^4.17.21",
     "@types/pako": "^2.0.3",
     "mini-svg-data-uri": "^1.4.4",
-    "prettier-plugin-tailwindcss": "^0.5.12",
-    "vconsole": "^3.15.1"
+    "prettier-plugin-tailwindcss": "^0.5.12"
   },
   "workspaces": {
     "packages": [
