+++ packages/smartsale-db/src/client.ts
@@ -2,4 +2,4 @@ import { PrismaClient } from '../prisma/client'
 
 export const db = new PrismaClient({})
 
-// 97
+// 98
